CSS: Changed the animation fade from the back-ground color of the header to the color of the header.
Having the background color of the header looked kinda hideous so instead, I changed the animation fade in / out from the background color to the font color and also changed the original font color from black to orange so it looks like its fading in from nothing and into the view. CSS is pretty fun!