Resolved the mobile responsiveness bug on the updates page
I couldn't figure out why the buttons on the updates page weren't mobile responsive. I tested if it was the button, changed the css, and also put it in a div. I finally figured it out by copying the index.html page and calling it test.html. The index page and the test page were both mobile responsive but the updates page wasn't. I looked at the head tag for the index.html and test.html and noticed that they both had a viewport meta tag - this tag was missing from the updates.html page. I have added it in and deleted the test.html page. This should do the trick. Let's try it out.