(1) Added counter that countdowns 2 minutes and (2) formatted text and button positioning
(1) Added a countdown timer that pops up a 'times up' when the time is reached. Also added a stop watch counter to it.

(2) Formatting
a. Made Bold: Mango Juice Money Counter and Sell Mango Juice Button

b. Moved mango juice money counter a little lower so it doesn't touch up against the top of the screen